test/netbench.cc: Handle the final responses before exiting
Knowing whether we've processed all the packets requires handling the responses,
which introduces another atomic.  In the future, we might want to boost the
throughput by using a separate counter for each core (since Mica guarantees that
a handler will only be invoked on the thread that made its request).  If we
decided we don't care about sample ordering, we could also split the buffer into
per-thread chunks and use a separate kId counter for each thread as well.